+++ ansible_rulebook/rule_set_runner.py
@@ -33,6 +33,7 @@ from ansible_rulebook.action.control import Control
 from ansible_rulebook.action.debug import Debug
 from ansible_rulebook.action.metadata import Metadata
 from ansible_rulebook.action.noop import Noop
+from ansible_rulebook.action.pg_notify import PGNotify
 from ansible_rulebook.action.post_event import PostEvent
 from ansible_rulebook.action.print_event import PrintEvent
 from ansible_rulebook.action.retract_fact import RetractFact
@@ -75,6 +76,7 @@ ACTION_CLASSES = {
     "run_module": RunModule,
     "run_job_template": RunJobTemplate,
     "run_workflow_template": RunWorkflowTemplate,
+    "pg_notify": PGNotify,
 }