+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Support for standalone boolean in conditions
 - Add basic auth to controller
 
+### Changed
+- Generic print as well as printing of events use new banner style
+
 ### Fixed
 
 ## [1.0.4] - 2023-10-30