+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Generic print as well as printing of events use new banner style
 
 ### Fixed
+- Support a base path for the controller url
 
 ## [1.0.4] - 2023-10-30